make the flip animation fit with the drop animation
When flipping a note to get a new one, afterwards the tiddler drops
from the top. To be consistent with the idea of a new note, we should cause it
to move to the top of the screen whilst being flipped, this will then merge into
the drop animation and be more seamless.