keep track of notes with user entered titles
this makes sure any note where the title has been set by the user
has a field _title_validated set.

the idea is that we can use this field to know when a note is allowed
to be synced.

Ben Gillies chrjs.store may have a way of handling this better.