when save to server fails display warning message instead of error
the save to server could be caused by a bad internet connection.
We don't want the user to worry about having no internet connection
so that the tool works offline.

This resets the ui allowing a user to write a new note despite
lack of internet connection but maintains it in the list of pending
tiddlers to be synced.