add test to check slow validation of title
sometimes it will take time for the server to tell the notes app
that a note already exists with a given title and cannot be used.

In the current setup if a user sets a title then quickly hits refresh
before the title can be validated, the title they have inputted is lost

since the validation could be slow due to lack of internet connection we must
make sure that in these cases the title is retained - just not validated.