when preloading existing notes from cache use hash rather than pushState
Since one of the primary goals of the app is for it to work offline and the
fact that cache manifests work by listing known files we shouldn't use
pushState as we are currently doing.

When encountering a url /takenote/tiddler/foo it will first look in the cache
manifest and not find it, thus will then request the resource from the server
even though the resource is the same.

Using a hash instead overcomes this problem. These tests reflect that change.

There is no problem with using a hash as we do not need search engine optimisation
for the authoring side of this app.