validate note titles during sync
this makes use of validateNote to check the title of each note to
be synced. If there is no note on the server with that name it will
be synced, however if there is it will be left so the user can rename
it.

all user feedback is now done in the giveFeedback function

all tests now pass in the test suite and issue #28 should be addressed