remember declines to geo location
some browsers are bad and do not remember the users preference on
geo location meaning at the moment everytime the app is loaded the user
is asked if they want to share their location.

This change adds a config (json) to localStorage. noGeoTiddlers is set
to the current date if the user says no (watchPosition error handler thrown)

when loading the config we check this setting and if set check if it is more
than a day old. If so we flush it.

This means user will only requested once in a day

TODO: allow user to toggle this setting during the day