update recent items if an internet connection is available
if tiddlers are deleted or edited in tiddlywiki they do not appear in
takenote. This corrects this by looking at the most recently edited tiddlers
on the server and updating the list.

note currently this means binary tiddlers will appear in the list of recently
edited tiddlers

this is a suggested implementation of #75