force links in the backstage to open in the same window
this means that when switching between the dashboard and the takenote
tool the window remains the same (the default behaviour for iphone/ipad
is to open a new safari window). This keeps the illusion of a single page
app whilst still allowing nice uris. It does this by overriding links in the
backstage (the black bar at the top of the page) and changing their behaviour